add Scenario Outline statuses to the Pickle reports (for each example) split JsonTable header and rows (to supports Status object in the json file) small Re-sharper refactoring
